Create trait_bonuses table
Bonuses are complicated; feats and traits can have multiple bonuses, which
can change depending on certain conditions. The simplest way to store all of
this is to have one "bonus" object for each different state of bonus that can
exist for a trait, then have the user select which bonuses should apply
during any given roll.