Remove bonuses from database
Keeping track of all the conditional bonuses in the game is an insane task;
I'm just going to do a simple text search in traits, feats, etc. for any given
roll. For example, if you roll damage, it'll search your traits and feats for
"damage" and show you any relevant ones for you to add in yourself.