Convert python data to json & add functionality
I've added the skills listing functionality, and converted the old script
skills into json format.

The character data is being represented in JSON format, and ideally I'd like
for the python script to be the front-end for the overall application, which
should be able to use either a database backend, or import characters directly
from files generated by the backend.