Add main loop and user input function
I'm probably going to end up ditching the idea for a database and just have
everything be json. Makes things a lot easier. With that in mind, this is
going to work like a dedicated program, written entirely in python (mostly as
a python learning experience). It's going to have a prompt within a main loop,
and will eventually handle generating and writing to character sheets.