Improve demo functionality of flask app
The intention here is to create the API using Flask, and then have a proper
front-end layer on top that makes requests to the Flask API endpoints. I'm
trying to work through the process of handling user data across requests; I
want to use a simple cookie, but I'm getting issues with handling them. For
now, the dev Flask setup is using sessions, but eventually I want to be able to
store the JSON in a cookie that a user can return to in a new browser session.