Change back to using cookies
There's a problem with my cookie plan: the character data is way too large
for a cookie. Like, my level 1 character is twice as large as the cookie limit
of 4KB. That means I need to use a different method of caching the user's
character sheet data, preferably all contained in the Flask app / javascript.

I'm currently looking into using local storage, but everything I read is
telling me that this means using javascript, which I was hoping to avoid for
now. I wanted to get an API working that basically uses client-side storage as
the data source, for two main reasons:

+ It keeps server-side stuff simple by avoiding logins, databases, caching, etc.
+ It keeps the user from losing their character sheet if they forget to export
their json

I know it's not exactly efficient to be sending the character data back and
forth for each API request, but I wanted this to be a simple self-hosted
solution that didn't require a login, internal or external, and caching user
data outside of a browser session seems to necessitate a login or server-side
storage.

So, my options:

+ Use local storage via javascript (this kinda breaks the decoupled API idea,
as the data would all be sourced from the javascript in the page instead of
from flask directly)
+ Ditch the idea and stick with browser sessions, as that's probably robust
enough for my needs