Change parser_edit flags
Before, all of the optional flags also had a default value of "". This
is dumb and doesn't make sense. Also, "--name" was optional, when
really we want it to be required. Now, we can just add each optional
argument to a dict if it exists, and pass the whole dict into
updateFeat.